+++ src/regdata.cpp
@@ -82,7 +82,7 @@ regdata::regdata(phedata &phed, gendata &gend, int snpnum,
     if (snpnum > 0)
         for (int j = 0; j < ngpreds; j++)
         {
-            float snpdata[nids];
+            double snpdata[nids];
             gend.get_var(snpnum * ngpreds + j, snpdata);
             for (int i = 0; i < nids; i++)
                 X.put(snpdata[i], i, (ncov - ngpreds + 1 + j));
@@ -97,7 +97,7 @@ void regdata::update_snp(gendata &gend, int snpnum)
 {
     for (int j = 0; j < ngpreds; j++)
     {
-        float snpdata[nids];
+        double snpdata[nids];
         for (int i = 0; i < nids; i++)
         {
             masked_data[i] = 0;
